# engine/coze_plugin/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Liuguang Video Engine",
    "description": "流光剪影 P0 引擎：草稿 CRUD + FFmpeg 云渲染 + TTS",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts": {
      "get": {
        "summary": "List Drafts",
        "operationId": "list_drafts_api_v1_drafts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Draft",
        "operationId": "create_draft_api_v1_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts/{draft_id}": {
      "get": {
        "summary": "Get Draft",
        "operationId": "get_draft_api_v1_drafts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Draft",
        "operationId": "delete_draft_api_v1_drafts__draft_id__delete",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts/{draft_id}/{resource}": {
      "post": {
        "summary": "Add Resource",
        "description": "添加素材/文字；body 支持单个对象或数组（批量）。",
        "operationId": "add_resource_api_v1_drafts__draft_id___resource__post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          },
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaBatchBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Add Resource Api V1 Drafts  Draft Id   Resource  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts/{draft_id}/{resource}/{segment_id}": {
      "patch": {
        "summary": "Patch Resource",
        "operationId": "patch_resource_api_v1_drafts__draft_id___resource___segment_id__patch",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          },
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource"
            }
          },
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Segment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSegReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove Resource",
        "operationId": "remove_resource_api_v1_drafts__draft_id___resource___segment_id__delete",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          },
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource"
            }
          },
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Segment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/probe": {
      "get": {
        "summary": "Media Probe",
        "operationId": "media_probe_api_v1_media_probe_get",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/tts": {
      "post": {
        "summary": "Ai Tts",
        "operationId": "ai_tts_api_v1_ai_tts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/render/tasks": {
      "post": {
        "summary": "Create Render Task",
        "description": "P0 同步渲染：请求阻塞至渲染完成（小样场景）。P1 迁移到任务总线异步。",
        "operationId": "create_render_task_api_v1_render_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/render/tasks/{task_id}": {
      "get": {
        "summary": "Get Render Task",
        "operationId": "get_render_task_api_v1_render_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddMediaReq": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "start": {
            "type": "number",
            "minimum": 0.0,
            "title": "Start",
            "default": 0.0
          },
          "end": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "source_in": {
            "type": "number",
            "minimum": 0.0,
            "title": "Source In",
            "default": 0.0
          },
          "level": {
            "type": "integer",
            "title": "Level",
            "default": 0
          },
          "x": {
            "type": "integer",
            "title": "X",
            "default": 0
          },
          "y": {
            "type": "integer",
            "title": "Y",
            "default": 0
          },
          "scale": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Scale",
            "default": 1.0
          },
          "alpha": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Alpha",
            "default": 1.0
          },
          "volume": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Volume",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "AddMediaReq",
        "description": "video/image/audio 通用添加参数。"
      },
      "AddTextReq": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "start": {
            "type": "number",
            "minimum": 0.0,
            "title": "Start",
            "default": 0.0
          },
          "end": {
            "type": "number",
            "title": "End"
          },
          "font_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Font Size",
            "default": 64
          },
          "font_color": {
            "type": "string",
            "title": "Font Color",
            "default": "white"
          },
          "align": {
            "type": "string",
            "enum": [
              "left",
              "center",
              "right"
            ],
            "title": "Align",
            "default": "center"
          },
          "border_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Border Color"
          },
          "border_width": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Border Width",
            "default": 0
          },
          "x": {
            "type": "integer",
            "title": "X",
            "default": 0
          },
          "y": {
            "type": "integer",
            "title": "Y",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "text",
          "end"
        ],
        "title": "AddTextReq"
      },
      "CreateDraftReq": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "未命名草稿"
          },
          "width": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Width",
            "default": 1080
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Height",
            "default": 1920
          },
          "fps": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Fps",
            "default": 30
          }
        },
        "type": "object",
        "title": "CreateDraftReq"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MediaBatchBody": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/AddMediaReq"
          },
          {
            "$ref": "#/components/schemas/AddTextReq"
          },
          {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AddMediaReq"
                },
                {
                  "$ref": "#/components/schemas/AddTextReq"
                }
              ]
            },
            "type": "array"
          }
        ],
        "title": "MediaBatchBody",
        "description": "单个素材/文字对象，或其数组（批量）。"
      },
      "PatchSegReq": {
        "properties": {
          "start": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "source_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source In"
          },
          "x": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "scale": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale"
          },
          "alpha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alpha"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "font_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Font Size"
          },
          "font_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Font Color"
          },
          "align": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "left",
                  "center",
                  "right"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Align"
          }
        },
        "type": "object",
        "title": "PatchSegReq"
      },
      "RenderReq": {
        "properties": {
          "draft_id": {
            "type": "string",
            "title": "Draft Id"
          },
          "crf": {
            "type": "integer",
            "maximum": 32.0,
            "minimum": 14.0,
            "title": "Crf",
            "default": 20
          },
          "preset": {
            "type": "string",
            "enum": [
              "ultrafast",
              "veryfast",
              "fast",
              "medium"
            ],
            "title": "Preset",
            "default": "veryfast"
          }
        },
        "type": "object",
        "required": [
          "draft_id"
        ],
        "title": "RenderReq"
      },
      "TTSReq": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          },
          "engine": {
            "type": "string",
            "enum": [
              "edge_tts",
              "mac_say"
            ],
            "title": "Engine",
            "default": "edge_tts"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TTSReq"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}